Much reduced errors.
Had to modify the submission steps to properly create order/submission.
Need to have order and submission IO models.
Allow belongs_to association to have :through specified.
Needed has_many :submissions, :through => :orders on Project
Removed project/study from Request I/O & endpoints.
Removed project/study from includes in various places.
Moved QuotaException into a sensible place.
Different behaviour for quota checking means different scenarios.
Asset group names come from the order UUID now.
Removed numericality check off read_length due to a Rails issue.

Seems Rails only adds one error per field on an association, so having
the numericality validation on read_length means that you only ever see
it, and not the inclusion error that might be present.  It doesn't make
sense to validate the numericality, given that the list itself is
numeric.
