+++ app/api/endpoints/requests.rb
@@ -4,8 +4,6 @@ class ::Endpoints::Requests < ::Core::Endpoint::Base
   end
 
   instance do
-    belongs_to(:study,        :json => 'study')
-    belongs_to(:project,      :json => 'project')
     belongs_to(:asset,        :json => 'source_asset')
     belongs_to(:target_asset, :json => 'target_asset')
     belongs_to(:submission,   :json => 'submission')
