+++ app/api/endpoints/submission_templates.rb
@@ -6,9 +6,9 @@ class Endpoints::SubmissionTemplates < Core::Endpoint::Base
   instance do
     factory(:to => 'submissions', :json => 'submissions') do |request, _|
       ActiveRecord::Base.transaction do
-        attributes = ::Io::Submission.map_parameters_to_attributes(request.json)
+        attributes = ::Io::Submission.map_parameters_to_attributes(request.json)[:order_attributes]
         attributes[:request_options] = attributes.delete(:request_options_structured)
-        request.create!(attributes.merge(:user => request.user))
+        request.target.create_with_submission!(attributes.merge(:user => request.user)).submission
       end
     end
   end
