--- app/models/exception/quota_exception.rb
@@ -1,2 +1,0 @@
-class QuotaException < Exception
-end
