+++ app/models/submission/workflow.rb
@@ -1,5 +1,3 @@
-require 'exception/quota_exception'
-
 class Submission::Workflow < ActiveRecord::Base
   has_many :request_types
   has_many :orders
