+++ db/schema.rb
@@ -9,7 +9,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20111010144613) do
+ActiveRecord::Schema.define(:version => 20111014114524) do
 
   create_table "aliquots", :force => true do |t|
     t.integer  "receptacle_id",    :null => false
